feat: refactor so all users benefit
This pushes the optimisation into BranchTable() so that all users of
that function benefit.
